requirements.txt
azure-core==1.32.0
azure-identity==1.19.0
azure-storage-blob==12.24.0
azure-storage-file-datalake==12.18.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorama==0.4.6
cryptography==43.0.3
idna==3.10
isodate==0.7.2
msal==1.31.1
msal-extensions==1.2.0
portalocker==2.10.1
pycparser==2.22
PyJWT==2.10.0
pywin32==308
requests==2.32.3
six==1.16.0
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
